requirements.txt
requests==2.27.1
beautifulsoup4==4.10.0
python-dotenv==0.20.0
PyMySQL==1.0.2
pdfminer3k==1.3.4
Pillow==9.1.0